fix: typo (#27)
or this could better read like so (?)

```
There is an issue related to the complexity of understanding the scenario to accurately create metrics. This problem might be resolved by updating the Configuration diagrams. The update should include defining responsibilities for generating a Metrics Plan and creating a separate endpoint to add metrics according to this plan. Implementing these changes for a specific service is crucial; otherwise, it will become computationally burdensome to update metrics within the overall configuration.
```